xcb: support receiving events from foreign windows
This is needed so that Qt can be aware of a window created and managed by
X11WS. Otherwise it could wrongly assume that the app is entirely
unfocused.

We manually set the event mask on Qt's xcb connection. This patch allows
app code to call addWindowEventListener to have the event processed by
the QXcb* machinery.