qwindowswindow: hack to improve support for "quasi-frameless" windows
The application code is free to override WM_NCCALCSIZE to resize the
client area (such as to include the titlebar). When the window is
created, Qt is not aware of that adjustment and would create an
oversized window, accounting for a title-bar that will not be there.

This commit introduces an ugly hack. Setting the following property:
_q_windowsClientAreaIncludesTitlebar
will cause the window creation code to temporarily use "custom margins"
which result in extending the client area into the titlebar. Note that
setting the same "custom margins" value from application code would have
other unintended consequences and so it wouldn't be a viable solution.